Abstract main Events component into separate folder/files
