+++ src/applications/static-pages/events/components/Events/constants.js
@@ -1,0 +1,1 @@
+export const perPage = 10;
