+++ src/applications/static-pages/events/components/Events/constants.unit.spec.js
@@ -1,0 +1,11 @@
+// Node modules.
+import { expect } from 'chai';
+// Relative imports.
+import { perPage } from './constants';
+
+describe('Events constants.js', () => {
+  it('perPage', () => {
+    // Assertions.
+    expect(perPage).to.equal(10);
+  });
+});
